// src/app/models/car.ts
export interface Car {

    id: number;
    description: string;
    address: string;
    brand: string;
    model: string;
    modelYear: number;
    issuance: string;
    mileage: string;
    fuel: string;
    color: string;
    numberDoors: number;
    horsPower: number;
    price: number;
    picture: String;
    solde?: boolean;
}